Refactor FixedAssetController and import scripts for robustness
The commit includes revisions of the FixedAssetController and two import scripts, namely CapexImport.php and MasterlistImport.php. In FixedAssetController.php, a null-coalescing operator was added to prevent undefined property errors if 'sub_capex' property does not exist.

In CapexImport.php and MasterlistImport.php files, the '->withTrashed()' method was introduced to queries so the application can also retrieve 'soft-deleted' capex entries. Additionally, vladimirTagGenerator function was refactored in the MasterlistImport.php file to improve the handling of EAN-13 (European Article Numbering) code generation and checking for duplicates.