"Refactor and optimise code to improve structure and maintenance
The 'telescope.php' configuration file was removed as the Telescope package is unnecessary in the current build. The 'TypeOfRequestController.php' was moved to a new directory 'MasterList' for better structuring and organisation of the codebase. Changes were also made in 'api.php' to update the path to the relocated controller file. This update will make the code easier to navigate and maintain."