Add Movement Status functionality
Added the 'MovementStatus' functionality for tracking the status of asset movements. This includes the addition of new 'MovementStatus' models, HTTP Requests, and Controllers. Beside this, the backend validation was also implemented for status changes during Post, Put, and Patch requests. Corresponding migration for Movement Status table was created in the database schema. This new feature will help in managing the movement statuses of fixed assets in a more systematic and efficient way.