Refactor FixedAssetRequest, add SubCapex and remove project_name
The commit includes a restructuring of the FixedAssetRequest class in order to refactor and improve code clarity. SubCapex model has been added and instances where 'project_name' was used have been replaced with 'capex_id' or 'sub_capex_id' as a more reliable and consistent reference. Also, changes were made in some validation rules to ensure accurate data entry. Now, depreciate and status attributes are validated by checking if they exist in corresponding tables. This was done to conform to the database schema change and to make the code more robust.