Revamp validation rules for accountable field
The validation rules for the 'accountable' field in the FixedAssetRequest, FixedAssetUpdateRequest files have been updated. Previously, failure was triggered if the accountable person's full name was missing, but now the rule is more comprehensive. Now, it checks if the accountable person's full name exists before proceeding with the validation. If it does not, a failure message is returned. Furthermore, if the accountable field is null, the validation simply continues, offering more robust functionality. Also, the error message 'faStatus.in' was unnecessarily present in the FixedAssetRequest file and has been removed.