Update accountable field validation in MasterlistImport
The validation rules for the 'accountable' field were improved. Now it takes into account the type of 'accountability' to determine if the field is required or not. This prevents unwanted gaps in the data and ensures accuracy when 'accountability' is of type 'Common'. A default value was also added when the field data is not available.