Update validation rule for type_of_request and capex
The validation rule for 'type_of_request' field has been updated to check from a dynamic list instead of static one. This was done by changing 'in:Asset,CAPEX...' to 'exists:type_of_requests,type_of_request_name'. In addition, a new validation condition is added to ensure that capex and sub_capex fields are empty when the 'type_of_request' is not CAPEX.