Refactor and improve code readability in FixedAssetRequest
This commit refactors the FixedAssetRequest class for the sake of readability and code quality. Space around the conditional control structures and functions have been made consistent according to the PSR-2 coding style guide for PHP. The 'TypeOfRequest' model was also imported but not used in this commit, indicating that it may be used in future updates.