Refactor asset data retrieval for accuracy
The asset data retrieval logic was updated to ensure appropriate mapping and accuracy of the data. The divison, company, and location id's are now retrieved from the department object instead of the asset directly. This adjustment was necessary because the previous version could introduce errors as not all fields were pertaining to the correct parent field. Furthermore, the 'est_useful_life' field was moved to be a common field rather than a major category specific field, offers more streamlined access for multiple categories. Lastly, the 'salvage_value' field was uncommented to make it available for processing.