Update implementation of tag number generator
The tag number generator in MasterlistImport and FixedAssetController has been moved to a repository. The reason is to increase the testability by decoupling the generation logic from the controllers and import jobs. It also makes it possible to reuse the tag number generation in other parts of the application without duplicating the code.

The update also includes changes in method calls and constructor to use the new repository. Moreover, unneeded code pieces and some unused validations have been removed.

This commit doesn't affect the working logic of the application but just a clean-up for the improvement of the code quality and maintainability.