Refactor code for readability, rename method and add new transformation method
Removed unnecessary comments and extra lines in VladimirTagGeneratorRepository for better readability. In FixedAssetRepository, redundant code was removed, transformFixedAsset method renamed to transformSingleFixedAsset and a new method, transformFixedAsset, was added. This new method uses transformSingleFixedAsset to transform each asset in a collection, making the transformation process clearer and more reusable. Changes in function calls were done in FixedAssetController as a result of the method being renamed.