Refactor error messages and add CapexExport function
In this commit, the error messages in SubCapexController and CapexController were updated for better readability. A capexExport function was also added to allow exporting of Capex with search, filter by date and status features. Moreover, the export method is added to api routes. In MajorCategoryController, logic to adjust the end_depreciation when updating a major_category is implemented. Some unnecessary white-spaces in MasterlistExportController were removed.