"Add validation messages, refactor IP format validation"
In this commit, predefined validation messages have been added for the PrinterIPRequest. This allows for better feedback about the specific requirement that failed during validation.

In addition, the IP format validation that was previously in the PrinterIPController has been removed. This change was made because it was considered better to do this validation at the request level, keeping the controller simpler and following better practices of validation.