Remove division_id validation and use accountable full name
Removed the unnecessary validation for division_id within the store and update methods of FixedAssetController. This validation was redundant and causing confusion in case implemented as the division_id is not required for FixedAsset creation. Also, merged the 'accountable' value from the request to use the full name of the accountability from the request data, providing more clarity and meaningful data in the system.