Refactor FixedAsset modules for efficiency
The update and creation methods in FixedAsset modules were refactored to improve the accuracy and speed of the application. The changes took account of potential null variables and made use of the request() helper function instead of instantiating a new request object. Also, The changes include simplifying validation rules and minimizing data manipulation for the accountable field from FixedAssetUpdateRequest. In the FixedAssetRepository file, the constructor's visibility was changed to protected level and the use of update method instead of create was put in place to make sure the changes affect the existing record only. Other changes include fixing indentation and removing unnecessary comments making the code cleaner and easier to understand. Lastly, the access level of few class properties were changed from private to protected to allow inheritance and improve encapsulation in Masterlist/FixedAssetController.