Refactor division references and fix minor typos
This commit addresses some changes in the MasterlistExportController and MasterlistImport files. Primarily, it refines how the app references divisions - instead of attempting to reach it directly, it does so through a related department attribute. This ensures we still get the division_name even if the division object is null.

Though a minor change, it adds up to the overall user experience as it helps prevent possible crashes and improves data consistency.

Furthermore, some typos noticed in the comments in the MasterlistImport class have been fixed for clarity and readability. Recommended functions are also added to retrieve and find employee data.