"Normalize depreciation method & optimize code"
The code for saving and updating 'depreciation_method' in the databases was modified to convert to uppercase if the method is 'STL', and to convert to title case otherwise. This change was made to normalize the method's format in the AdditionalCostRepository, FixedAssetRepository, and MajorCategoryController. Code optimization was also performed in MajorCategoryController by creating a new 'applyEndDepreciation' function to eliminate duplicate code. In FixedAssetRepository, ‘searchInMainAttributes’ and ‘searchInRelationAttributes’ functions were created to avoid repeating loops within the 'search' logic. Additionally, an 'additional_cost_count' attribute was added to provide count information of additional costs for each fixed asset. Lastly, the validation rules for an 'AdditionalCostRequest' have been updated with additional conditions.