Refactor fixed asset component for status property
This commit addresses several changes in the fixed asset module, primarily related to the introduction of a status property. The update() function in FixedAssetController is now expanded with an 'archived()' method, enabling assets to be deactivated with a new 'status' attribute. If the status is changed, additional checks are made to confirm the association with SubCapex, minor categories and type of requests. If these checks do not pass, an error is thrown.

In addition, search functionality now includes a status parameter, allowing assets to be filtered by their activation status. The calculateRemainingBookValue() function in FixedAssetExportRepository has also been refactored to exclude the 'monthly_depreciation_rate'.

Various other minor changes include amendments in rules validation and strings formatting. For instance, responsible parties ('accountable') are now saved in upper case, and 'remarks' attribute is limited to 255 characters.