Refine logic and improve data validation of various methods
This commit refines logic and improves data validation for various methods throughout the application.

In PrinterIPController.php, the activateIP method was modified to deactivate the IP if it is already active and activate it when it is inactive as well as remove some old commented-out code. The LocationController was also updated to check if the department table is empty and return an error message before proceeding with the code.

Changes were also made in FixedAssetExportRepository.php and DepartmentController.php to improve error handling by changing the response message to 'No data found' when no results are obtained from the database.

A validation was added in CapexImport.php to verify the sub capex format. The update also includes improved data validation in MasterlistImport.php by passing an empty array if the collection is empty. Minor changes were made to other parts of the code as well for code cleanliness and readability.