Remove unused function and refactor data sync checks
Removed the destroy function from the CompanyController as it is not being used. Additionally, added checks to see if the department in LocationController and company in DepartmentController exist before data sync. The checks were moved to top to avoid unnecessary processing of request data, especially when the pre-requisites are not met. It also prevents the creation of entities with unmet dependencies.

This is more efficient and ensures robustness in data synchronisation operations.