Add account title syncing to MinorCategory
MinorCategory now has the ability to be synced with the account title. Changes include storing the account_title_sync_id when creating or updating a MinorCategory. The account title details are also returned in the response. Additionally, a new validation rule has been added to ensure the uniqueness of the combination of major category, minor category, and account title sync id. The addition of account title syncing enhances the customization of MinorCategory and provides more detail in responses.