"Add a department_location pivot table and refactor relationships
Created a new 'department_location' pivot table in the database to serve as a bridge for the many-to-many relationship between Departments and Locations. Refactored the 'departments' and 'locations' tables by removing the direct foreign key from 'departments' to 'locations' and updating the indexes on 'sync_id' for both tables. Also, adjusted the Department and Location Eloquent models to now use the new pivot table for defining their relationship. This change allows for more flexibility when assigning departments to different locations."