Update depreciation calculation from depreciable basis to acquisition cost
The depreciation calculations previously used the depreciable basis for calculating depreciation amounts. This commit changes these calculations to use the acquisition cost instead. This change is made to align our asset depreciation calculations with standard accounting practices. By using the acquisition cost, the system is now able to provide a better representation of asset value decrease over its useful life. This impacts the calculation of monthly and yearly depreciation, as well as the remaining book value of the asset.