Update validation error messages and asset tag check
This commit updates the error messages in AdditionalCostImport.php to accurately reflect when a combination of department and company is invalid, as the location was not actually part of this check. Additionally, a comment which was no longer relevant has been removed from MasterlistImport.php. The check to define an old asset in the MasterlistImport.php has been corrected, as it had a typographical error in the key it was checking in the collection.