"Comment out voucher validation in AdditionalCostImport.php
The voucher validation function in the AdditionalCostImport.php file has been commented out. This change is temporary and done for debugging purposes for an issue related to the validation of vouchers. Now, the application will not validate the uniqueness of the fixed assets' vouchers. A permanent solution would replace this temporary hack after debugging is complete."