Add add_cost_sequence to FixedAssetExport and prevent division by zero
Two changes were made in this commit. Firstly, an additional field "add_cost_sequence" was added to the export functionality in FixedAssetExportRepository.php to enhance the data details available for export.

Secondly, a check has been added in the CalculationRepository.php to prevent a division by zero error when both acquisition cost and scrap value are zero. This would mitigate potential crashes in the depreciation calculations.