"Refine voucher validation and calculate depreciation accurately
Updated the voucher validation in multiple request files to include a case when voucher value equals to '-'. This change will ensure a voucher is provided whenever it is required by the depreciation status.

The CalculationRepository's getAccumulatedCost and getRemainingBookValue methods are also updated to account for cases when the accumulated cost is greater than the depreciable basis and when the remaining book value is less than zero, respectively. This ensures that the calculated depreciation values are accurate.

FixedAssetController, AdditionalCostController and FixedAssetExportRepository have been modified to use the new calculation method signature introducing the depreciable basis argument. The purpose is to provide accurate calculation for accumulated cost."