Update department validation in asset export and import
Modified how department validation is done in FixedAssetExportRepository, AdditionalCostImport and MasterlistImport. Previously, validation between the department and company was done directly, without considering the active status of the department. The new implementation breaks the department validation into two parts. First, it checks if department exists, is active and corresponds to the designated department name. Secondly, it verifies the combination of department and company. This change ensures that only departments that are currently active and have a valid association with the company are passed by the validation process.