"Add Approver and Approvers Assigning functionality
This commit includes new logic to manage 'approvers' and 'approvers assignment' in the system. Migration, Model, Request, Route, and Controller are created for 'approvers'. This allows admin users to add approvers who can approve certain actions in the system. Also, the ability to assign these approvers to specific users is added. In 'LocationController', extra information about 'department status' is included in the response."