Update voucher date for Fixed Asset and AdditionalCost
Voucher date has been added to both the Fixed Asset and Additional Cost models. This field has been added to the corresponding request files for validation purposes. In addition, it has also been modified in the repository classes. This change arises from the need to specify the date of the voucher. This information is vital for processes such as depreciation calculation and more. Validation of the voucher date ensures that the provided date is valid and in an expected format. The verification check also ensures that the date provided aligns properly with other entries in the database with the same voucher number. Furthermore, the check ensures that the voucher date is required when the depreciation status is either running or fully depreciated.