"Renamed AssigningApprover classes to AssignApprover and allowed null values for voucher dates
In this commit, the AssigningApprover Controller and Request classes were renamed to AssignApprover to better reflect their function. The routes and method names were updated correspondingly. In addition, certain fields in the AdditionalCost, FixedAsset, and their relative repositories were updated to allow null values for voucher dates, providing flexibility when data is not available. 'Supplier's Rebase' depreciation method in voucher related validation rules was added to AdditionalCost and FixedAsset requests to better align with the business rules. Furthermore, the maintenance command 'sanctum:prune-expired' is added to Kernel to automatically remove expired tokens daily."