Update error message in AuthController
The error message in the authentication controller was updated for cases when a user attempts to log in with an invalid username. Previously, the message clarified only that the username was not found. Now, it provides more generic feedback, informing that either the username and/or password is incorrect. This is a routine practice to prevent potential attackers from verifying whether a username exists, enhancing security.