Add AssetApprovalLoggerController and related routes, update existing controllers
The AssetApprovalLoggerController is introduced to handle the listing of resource activities. The dynamicPaginate function is used for paginated display on frontend. Corresponding routing is added in api.php. A new attribute, employee_id, is included in ApprovedRequestRepository for the convenience of tracking approvers. AssetApprovalController and its associated methods are reorganized for better code readability and performance. The changes aim to improve the handling of asset approval requests and their subsequent tracking for clearer audit trails.