Refactor error messages and status update logic
Changes were made in the error messages returned by the CreateAssetApprovalRequest to provide clearer instructions. The logic for status update in the ApprovedRequestRepository was also refactored changing the 'status' parameter to 'layer'. This was done to provide more control and granularity in the validation process. A 'updateToNull' function has also been added to reset the status of the approval requests.