Add department and subunit fields to User model
In this commit, the "Users" table in our database and the User model have been updated to include two additional fields - 'department_name' and 'subunit_name'. This helps in accurately categorizing users according to their department and subunit. The UserController and UserRequest files have been modified to accept and validate these new fields. Additionally, the Department model has been updated to include a one-to-many relationship function with the SubUnit model. A typo in the composer.json file has also been corrected. The routes file has been updated to include routing for the DepartmentUnitApproversController and SubUnitController. Lastly, unused commented code was cleaned up from the DivisionController file.