Add nullable 'sub_unit_code' field in 'sub_units' table
This commit introduces a new 'sub_unit_code' field in the 'sub_units' table. This field is intended to store unique codes for each sub_unit which can be null. This new feature is added considering the future requirement and the ability to store additional unique identifiers for sub_units if necessary.