Update success message and add validation for duplicate subunits
The success message in 'DepartmentUnitApproversController' was updated to be generic for better reusability. Additionally, a validation rule was added in 'CreateDepartmentUnitApproversRequest' to prevent the creation of duplicate approvers for the same subunit in a department. This ensures that there won't be any confusion in the assignation of approvers to subunits.